+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - n/a
 
+## [2.52.1] - 2025-12-11
+
+### Fixed
+
+- Import of `broken_rpmlib` in `KojiSource`
+
+## [2.52.0] - 2025-12-11
+
+### Added
+
+- Support for "alias signing key" in `ErrataSource` and `KojiSource`
+
 ## [2.51.2] - 2025-10-02
 
 ### Fixed
@@ -582,7 +594,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial stable release of project
 
-[Unreleased]: https://github.com/release-engineering/pushsource/compare/v2.51.2...HEAD
+[Unreleased]: https://github.com/release-engineering/pushsource/compare/v2.52.1...HEAD
+[2.52.1]: https://github.com/release-engineering/pushsource/compare/v2.52.0...v2.52.1
+[2.52.0]: https://github.com/release-engineering/pushsource/compare/v2.51.2...v2.52.0
 [2.51.2]: https://github.com/release-engineering/pushsource/compare/v2.51.1...v2.51.2
 [2.51.1]: https://github.com/release-engineering/pushsource/compare/v2.51.0...v2.51.1
 [2.51.0]: https://github.com/release-engineering/pushsource/compare/v2.50.0...v2.51.0